refactor : 주문 생성 및 주문번호 시퀀스 적용 수정 JwtAuthenticationFilter CORS 설정 주석 추가, 확인 후 삭제 또는 유지

+++ src/main/java/com/bootcamp/paymentdemo/domain/order/entity/ProductOrder.java
@@ -24,7 +24,7 @@ public class ProductOrder extends Base {
     private String nameSnap;
 
     @Column(nullable = false)
-    private Integer priceSnap;
+    private Long priceSnap;
 
     @Column(nullable = false)
     private Integer quantity;
@@ -46,7 +46,7 @@ public class ProductOrder extends Base {
             Product product,
             Order order,
             String nameSnap,
-            Integer priceSnap,
+            Long priceSnap,
             Integer quantity
     ) {
         ProductOrder productOrder = new ProductOrder();
